@@ -2,13 +2,14 @@ package test.steps;
 
 import cucumber.api.java.en.Then;
 import main.poms.MyAccountPage;
-import org.junit.Assert;
+import org.junit.jupiter.api.Assertions;
+
 
 public class CompareSteps {
     private MyAccountPage myAccPage = new MyAccountPage();
 
     @Then("^Check if something$")
     public void checkIfSomething() {
-        Assert.assertTrue(myAccPage.returnBtn().isDisplayed());
+      Assertions.assertTrue(myAccPage.returnBtn().isDisplayed());
     }
 }
