@@ -1,4 +1,4 @@
-package test.runners;
+package runners;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
